Refactor directory structure: move molecule/ to core/, openbabeltools/ and rdkittools/ directories
