--- libcoffee/openbabeltools.py
@@ -1,1 +1,0 @@
-from libcoffee.molecule.openbabel.mol import PybelMol
