+++ libcoffee/openbabeltools/__init__.py
@@ -1,0 +1,3 @@
+from libcoffee.openbabeltools.mol import PybelMol
+
+__all__ = ["PybelMol"]
