+++ libcoffee/openbabeltools/_util.py
