+++ libcoffee/openbabeltools/mol.py
@@ -8,7 +8,7 @@ import numpy.typing as npt
 from openbabel import pybel
 from openbabel.openbabel import OBConversion
 
-from libcoffee.molecule.molbase import MolBase
+from libcoffee.core.molbase import MolBase
 
 from ._util import combine_two_mols
