+++ libcoffee/openbabeltools/test_openbabel_mol.py
