--- libcoffee/rdkittools.py
@@ -1,2 +1,0 @@
-from libcoffee.molecule.rdkit.mol import RDKitMol
-from libcoffee.molecule.rdkit.util import calc_similarities
