+++ libcoffee/rdkittools/__init__.py
@@ -1,0 +1,4 @@
+from libcoffee.rdkittools.mol import RDKitMol
+from libcoffee.rdkittools.util import calc_similarities
+
+__all__ = ["RDKitMol", "calc_similarities"]
