+++ libcoffee/rdkittools/test_rdkit_mol.py
